requirements.txt
boto3==1.35.18
Flask==3.0.3
gunicorn==23.0.0
MarkupSafe==2.1.5
numpy==1.26.3
okta_jwt_verifier==0.2.5
python-dotenv==1.0.1
requests_aws4auth==1.3.1
tiktoken==0.8.0
langchain-text-splitters==0.3.1

# Local Development Requirements
pre-commit==3.8.0
autopep8==2.3.1
pylint==3.3.1
